Fix compilation warnings
This fix is the first part of an effort to resolve all warnings that
appear when building OpenJCEPlus

+++ src/test/java/ibm/jceplus/junit/base/BaseTestHKDFInterop.java
@@ -116,9 +116,9 @@ public class BaseTestHKDFInterop extends BaseTestJunit5Interop {
             HKDFParameters hkdfParametersBC = new HKDFParameters(ikmArray, saltArray,
                     infoArray);
             hkdfBytesGeneratorBC.init(hkdfParametersBC);
-            byte[] okmBC = new byte[(int) okmLength];
+            byte[] okmBC = new byte[okmLength];
 
-            hkdfBytesGeneratorBC.generateBytes(okmBC, 0, (int) okmLength);
+            hkdfBytesGeneratorBC.generateBytes(okmBC, 0, okmLength);
 
             assertArrayEquals(calcOkmArray, okmBC, "OpenJCEPlus and BC results don't match");
         }
@@ -158,9 +158,9 @@ public class BaseTestHKDFInterop extends BaseTestJunit5Interop {
                 HKDFParameters hkdfParametersBC = new HKDFParameters(ikmArray, saltArray,
                         infoArray);
                 hkdfBytesGeneratorBC.init(hkdfParametersBC);
-                byte[] okmBC = new byte[(int) okmLength];
+                byte[] okmBC = new byte[okmLength];
 
-                hkdfBytesGeneratorBC.generateBytes(okmBC, 0, (int) okmLength);
+                hkdfBytesGeneratorBC.generateBytes(okmBC, 0, okmLength);
                 assertArrayEquals(calcOkmArray, okmBC, "OpenJCEPlus and BC results don't match");
             }
         } catch (NoSuchAlgorithmException e) {
